Small refactor around the naming of the query to come in line with the APIM project, and putting it in it's own namespace
