--- src/SFA.DAS.LevyTransferMatching/Application/Queries/GetAllPledgesQuery.cs
@@ -1,11 +1,0 @@
-﻿using MediatR;
-using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace SFA.DAS.LevyTransferMatching.Application.Queries
-{
-    public class GetAllPledgesQuery : IRequest<GetAllPledgesResult>
-    {
-    }
-}
