--- src/SFA.DAS.LevyTransferMatching/Application/Queries/GetAllPledgesResult.cs
@@ -1,12 +1,0 @@
-﻿using SFA.DAS.LevyTransferMatching.Models;
-using System.Collections.Generic;
-
-namespace SFA.DAS.LevyTransferMatching.Application.Queries
-{
-    public class GetAllPledgesResult : List<Pledge>
-    {
-        public GetAllPledgesResult(IEnumerable<Pledge> collection) : base(collection)
-        {
-        }
-    }
-}
