+++ src/SFA.DAS.LevyTransferMatching/Application/Queries/GetPledges/GetPledgesQuery.cs
@@ -1,0 +1,11 @@
+﻿using MediatR;
+using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace SFA.DAS.LevyTransferMatching.Application.Queries.GetPledges
+{
+    public class GetPledgesQuery : IRequest<GetPledgesResult>
+    {
+    }
+}
