+++ src/SFA.DAS.LevyTransferMatching/Application/Queries/GetPledges/GetPledgesResult.cs
@@ -1,0 +1,12 @@
+﻿using SFA.DAS.LevyTransferMatching.Models;
+using System.Collections.Generic;
+
+namespace SFA.DAS.LevyTransferMatching.Application.Queries.GetPledges
+{
+    public class GetPledgesResult : List<Pledge>
+    {
+        public GetPledgesResult(IEnumerable<Pledge> collection) : base(collection)
+        {
+        }
+    }
+}
